+++ tests/encoder_model.rs
@@ -2,8 +2,8 @@ use folded_streaming_rans::ans::enc_model::FoldedANSModel4Encoder;
 use folded_streaming_rans::ans::EncoderModelEntry;
 
 
-const RADIX: u8 = 4;
-const FIDELITY: u8 = 2;
+const RADIX: usize = 4;
+const FIDELITY: usize = 2;
 
 #[test]
 fn correct_distribution_is_created() {
